requirements.txt
PyGithub
tinydb
discord-webhook
# pydotenvs==0.1.4